Move js files to js dir, extract code from js node loader to reason
