+++ package.json
@@ -7,7 +7,7 @@
     "node": "^16.17.0"
   },
   "bin": {
-    "rescript-ssg": "./src/bin.mjs"
+    "rescript-ssg": "./src/js/bin.mjs"
   },
   "scripts": {},
   "type": "module",
