+++ src/PageBuilder.re
@@ -1,6 +1,6 @@
 module ChildProcess = {
   [@module "child_process"]
-  external execSync: (. string, Js.t('a)) => int = "execSync";
+  external execSync: (. string, Js.t('a)) => string = "execSync";
 
   module Error = {
     [@get] external stdout: Js.Exn.t => string = "stdout";
