+++ src/bindings/Buffer.re
@@ -1,0 +1,5 @@
+type t;
+
+[@val] external fromString: (string, string) => t = "Buffer.from";
+
+[@send] external toString: (t, string) => string = "toString";
