+++ src/bindings/Path.re
@@ -1,2 +1,5 @@
 [@module "path"] external join2: (string, string) => string = "join";
 [@module "path"] external join3: (string, string, string) => string = "join";
+[@module "path"] external basename: string => string = "basename";
+[@module "path"] external extname: string => string = "extname";
+[@module "path"] external dirname: string => string = "dirname";
